libraries/async-rabbitmq/requirements-editable.txt
pydantic-base-settings
